Added the basic javascript client and node.js server to the project.
index.html contains an HTML5 canvas which is drawn on by draw.js
draw.js also contains all the code for the layercake client,
except for two supporting classes which are defined in layerandsim.js

gameserver is basically just the chat server example from
miksago/node-websocket-server, with the supporting libs

at this point the server is useless, but the client can still play
offline. note that will attempt to load resources from uncc.ath.cx,
because that's where the old Java version that predated this resides.